proper rewrite of the minor mode I split the keymap into two part, the basic and advanced one (which can be toggled using a custom variable). Both part now feature a lot of comments specifying the use of each individual binding.
I also added a number of custom variables including their documentation at the beginning of the document enabling the user to navigate and delete faster.

The copy bindings are now residing at the *o* key, the yanking at *i*, and the indenting at *p*.

The first layer indenting C-p was broken and has been repaired.

In addition I added a bunch of if clauses checking whether a region was mark and a number of save-excursion environments