try to deal with the overwriting the loop adjusting the inheritance of the keymap variable were buggy and have been fixed.
The remaining problem is the following: lightning-keymap-mode uses the functions (current-local-map) and (current-minor-mode-maps) to determine all active maps, set a list of them as parent, and thus avoids to get distracted by override text properties. But those functions are only evaluated the moment lightning-keymap-mode is deactivated. Therefore they won't change when switching the buffer or/and major mode. This causes strong bugs like yas-snippet-expand is mapped on TAB all the time and there is almost no way to use helm or dired anymore.

To circumvent this problem I tried a hook which gets evaluated every time a major mode is changed, called *after-change-major-mode-hook*. But, as far as I understood the error messages, the bindings were just added on top of the ones already present. So I moved all the key bindings into a function generating lightning-keymap-mode-map. This way I can call it every time the hook function is evaluated and use a fresh version of the keymap.

Unfortunately something is massively wrong now and it doesn't work. There might be no other choice but to use either a different hook or some debugging on the hook function.