using post-command-hook There seems not to be an appropriate hook, which is run every time the user changes the focus to another buffer.
In principle I could trigger the reevaluation of the local-map and the minor-mode-map every time the user changes the buffer using the key bindings provided in the lightning-keymap-mode. But what happens if she uses C-x ARROW or the mouse instead? This might not be a good idea and does not generalize.

Instead I tried using the 'post-command-hook'. There, after every command run by the user of Emacs itself the keymap is reevaluated. This sounds extremely costy. But in the end there isn't a noticable loss of performance (at least not on the computer I'm currently working on).

Hooks like
 'buffer-list-update-hook' is a bad idea, because it is triggered even more often.

In order to reduce the number of debugging messages occuring using 'post-command-hook', I moved the definition of the lightning-debugging variable to the main body of the mode. Now the user just can execute (setq lightning-debugging 1) whereever she wants and active the debugging. 'nil' is deactivating it again.

But there are still a number of bugs to fix.