sparse evaluation & reinserting docs For some reason I accidentally removed the documentation. It's not present again.
I removed the import of the rect-mark package since I was not binding to it after all.

In order to save CPU time and not to slow down emacs there is now a global variable containing the major mode and a list of all active minor modes. Only if the content of this variable changes the overriding-local-map is reset.