global variable definition the `defvar` function is now used to initiate both the `lightning-keymap-mode-map` and `lightning-debugging` variables.
For the first one there is a significant change in this behaviour. Since the `lightning-keymap-mode-get-keymap` function also used the `defvar` function, the lightning keymap was NOT reset when calling it. Instead all the additional bindings were just appended. This was the reason for the "Wow, an indirect keymap entry!" error.

Now the `setq` function is used in the `lightning-keymap-mode-get-keymap` function. This way the variable is overwritten at every call."