inheritance exception + map refresh Sometime the `(current-minor-mode-maps)` functions returns several empty keymaps. An exception in the `if` clause has been added to cope this kind of behaviour.
As a result the final `list-of-all-active-maps` variable can be either a list or a keymap. Now its class is checked and the inheritance is set with or without the usage of the `make-composed-keymap` function concatenating the maps into a single one.

In addition a new instance of the `lightning-keymap-mode-map` variable is now created at the very beginning of the `lightning-keymap-post-command-function` function to assure the variable is properly set and an actual keymap