rewrote faster navigation In some minibuffers moving one line or one character does not make a lot of sense e.g. in `magit` buffers. Instead the faster fifth and sixth layer now also remap to <left>, <right>, <up>, and <down> multiple times.
To implement this quite a hack was necessary. The function `unread-command-events` takes a list of input events and executes them one after another.

The input to this function is generated by `make-list`, which takes an object and creates a list out of it holding *n* elements.

The object reproduced in the list is the keycode for <left>, <right> etc. E.g. ffor <left> the code is 10.

How to find out which code to take? I dont want to hard code it into the script. Instead do a string match to `key-translation-map`, which has the format *(10 . [left])*.

This implements the feature request in issue #4