updating string search I removed the dependency on `iedit` and instead bound a modified version of the `replace-string` function to the replacement keybindings. (line, paragraph, and buffer)
In addition I updated the cheatcheet (including the mappings of the j,k,l, and ; keys to the direction keys

The newly defined `lightning-keymap-mode-replace-string` function is a wrapper aroud the system function `replace-string`, but behaves somewhat different. The system function remembers the FROM and TO arguments of the last call and offers to do exactly the same thing the next time it is called by the user.

I found this behavior not to be of any particular use. Therefore I changed the default behavior by reading the word at point instead and offering this one as the default value for the FROM argument in the replace function.

Whether or not to use this new behavior is of course up to the user. By setting the `lightning-keymap-mode-modifies-string-replacement` to nil, the `string-replace` function without any modifications will be used instead

In addition I updated all corresponding documentation and added documentation at the beginning of the lightning-keymap-mode-functions.el and lightning-keymap-mode-modes.el files.